+++ requirements.txt
@@ -1,5 +1,6 @@
 scikit-learn>=0.22.1
 PyYAML>=5.1.2
+scipy>=1.5.4 # Avoid call deprecated np.typeDict
 numpy
 pytest
 toolwrapper
@@ -7,12 +8,11 @@ joblib
 sacremoses
 bicleaner-hardrules==2.5.1
 sentencepiece
-tensorflow>=2.6.5,<2.11
-markdown==3 # Temporary forced version until Transformers fix compatibility with tensorflow 2.11
+tensorflow>=2.8.4,<2.12,!=2.9
 bicleaner-ai-glove==0.2.1
 fuzzywuzzy
 python-Levenshtein
-transformers>=4.10.3,<4.25
+transformers>=4.26,<4.27
 huggingface-hub>=0.11,<0.12
 psutil
 regex
